move Table component to tables directory
