+++ shared/pages/Home/Orders/Orders.js
@@ -2,7 +2,7 @@ import React, { Component, Fragment } from 'react'
 import { connect } from 'redaction'
 
 import Row from './Row/Row'
-import Table from 'components/Table/Table'
+import Table from 'components/tables/Table/Table'
 import MyOrders from './MyOrders/MyOrders'
 import SearchSwap from 'components/SearchSwap/SearchSwap'
